migrate to shared logic in Detection class
